Extract even/odd checks into utility functions
